fix: Fix layouting on mobile
