+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml/Given_UIElement.cs
@@ -938,11 +938,7 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml
 
 			using var _ = new AssertionScope();
 
-#if __ANDROID__
-			ctl1.MeasureCount.Should().Be(2);
-#else
 			ctl1.MeasureCount.Should().Be(1);
-#endif
 			ctl2.MeasureCount.Should().Be(2);
 			ctl3.MeasureCount.Should().Be(1);
