+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/VirtualizingPanelLayout.Android.cs
@@ -1019,7 +1019,7 @@ namespace Microsoft.UI.Xaml.Controls
 
 			if (child.IsLayoutRequested || slotSize != previousAvailableSize)
 			{
-				var size = _layouter.MeasureChild(child, slotSize);
+				var size = MobileLayoutingHelpers.MeasureElement(child, slotSize);
 
 				if (ShouldApplyChildStretch)
 				{
@@ -1094,7 +1094,7 @@ namespace Microsoft.UI.Xaml.Controls
 				top = logicalBreadthOffset;
 			}
 			var frame = new global::Windows.Foundation.Rect(new global::Windows.Foundation.Point(left, top), size);
-			_layouter.ArrangeChild(child, frame);
+			MobileLayoutingHelpers.ArrangeElement(child, frame);
 
 			// Due to conversions between physical and logical coordinates, the actual child end can differ from the end we sent to the layouter by a little bit.
 			Debug.Assert(direction == GeneratorDirection.Forward || Math.Abs(GetChildEndWithMargin(child) - extentOffset) < 2, GetAssertMessage("Extent offset not applied correctly"));
