+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/VirtualizingPanelLayout.cs
@@ -33,10 +33,6 @@ namespace Microsoft.UI.Xaml.Controls
 		/// </summary>
 		/// <remarks>For <see cref="ItemsStackPanel"/> layouting this is identical to <see cref="Orientation"/> but for <see cref="ItemsWrapGrid"/> it is the opposite of <see cref="Orientation"/>.</remarks>
 		public abstract Orientation ScrollOrientation { get; }
-#if !UNO_REFERENCE_API
-		private protected readonly ILayouter _layouter = new VirtualizingPanelLayouter();
-		internal ILayouter Layouter => _layouter;
-#endif
 
 #pragma warning disable 67 // Unused member
 		[NotImplemented]
@@ -280,35 +276,6 @@ namespace Microsoft.UI.Xaml.Controls
 
 			return (minItem, min);
 		}
-
-#if !UNO_REFERENCE_API
-		private class VirtualizingPanelLayouter : Layouter
-		{
-
-			public VirtualizingPanelLayouter() : base(null)
-			{
-
-			}
-			protected override string Name => "VirtualizingPanelLayout";
-
-			protected override Size ArrangeOverride(Size finalSize)
-			{
-				throw new NotSupportedException($"{nameof(VirtualizingPanelLayouter)} is only used for measuring and arranging child views.");
-			}
-
-#if __ANDROID__
-			protected override void MeasureChild(Android.Views.View view, int widthSpec, int heightSpec)
-			{
-				view.Measure(widthSpec, heightSpec);
-			}
-#endif
-
-			protected override Size MeasureOverride(Size availableSize)
-			{
-				throw new NotSupportedException($"{nameof(VirtualizingPanelLayouter)} is only used for measuring and arranging child views.");
-			}
-		}
-#endif
 	}
 }
