+++ src/Uno.UI/UI/Xaml/Controls/TextBox/MultilineTextBoxView.iOS.cs
@@ -183,7 +183,7 @@ namespace Microsoft.UI.Xaml.Controls
 					expectedSize.Width = size.Width;//Take all the space given
 				}
 
-				var result = IFrameworkElementHelper.SizeThatFits(this, expectedSize);//Adjust for NaN and MaxHeight..
+				var result = SizeThatFits(expectedSize);//Adjust for NaN and MaxHeight..
 
 				return result;
 			}
