+++ src/Uno.UI/UI/Xaml/Controls/TextBox/SinglelineTextBoxView.iOS.cs
@@ -125,7 +125,7 @@ namespace Microsoft.UI.Xaml.Controls
 
 		public override CGSize SizeThatFits(CGSize size)
 		{
-			return IFrameworkElementHelper.SizeThatFits(this, base.SizeThatFits(size));
+			return base.SizeThatFits(size);
 		}
 
 		public override CGRect TextRect(CGRect forBounds)
