+++ src/Uno.UI/UI/Xaml/IFrameworkElement.cs
@@ -312,109 +312,6 @@ namespace Microsoft.UI.Xaml
 			return null;
 		}
 
-		public static CGSize Measure(this IFrameworkElement element, _Size availableSize)
-		{
-#if __IOS__ || __MACOS__
-			return ((View)element).SizeThatFits(new CoreGraphics.CGSize(availableSize.Width, availableSize.Height));
-#elif __ANDROID__
-			var widthSpec = ViewHelper.SpecFromLogicalSize(availableSize.Width);
-			var heightSpec = ViewHelper.SpecFromLogicalSize(availableSize.Height);
-
-			var view = ((View)element);
-			view.Measure(widthSpec, heightSpec);
-
-			return Uno.UI.Controls.BindableView.GetNativeMeasuredDimensionsFast(view)
-				.PhysicalToLogicalPixels();
-#else
-			return default(CGSize);
-#endif
-		}
-
-#if __MACOS__
-		public static CGSize SizeThatFits(this View element, _Size availableSize)
-		{
-			switch (element)
-			{
-				case NSControl nsControl:
-					return nsControl.SizeThatFits(availableSize);
-
-				case FrameworkElement fe:
-					{
-						fe.XamlMeasure(availableSize);
-						var desiredSize = fe.DesiredSize;
-						return new CGSize(desiredSize.Width, desiredSize.Height);
-					}
-
-				case IHasSizeThatFits scp:
-					return scp.SizeThatFits(availableSize);
-
-				case View nsview:
-					return nsview.FittingSize;
-
-				default:
-					throw new NotSupportedException($"Unsupported measure for {element}");
-			}
-		}
-
-#endif
-
-		public static CGSize SizeThatFits(IFrameworkElement e, CGSize size)
-		{
-			// Note that on iOS, the computation is intentionally kept as nfloat
-			// to handle discrepancies with the nfloat.NaN and double.NaN.
-
-			if (e.Visibility == Visibility.Collapsed)
-			{
-				return new CGSize(0, 0);
-			}
-
-			var (min, max) = e.GetMinMax();
-
-			var width = size
-				.Width
-				.NumberOrDefault(nfloat.PositiveInfinity)
-				.LocalMin((nfloat)max.Width)
-				.LocalMax((nfloat)min.Width);
-
-			var height = size
-				.Height
-				.NumberOrDefault(nfloat.PositiveInfinity)
-				.LocalMin((nfloat)max.Height)
-				.LocalMax((nfloat)min.Height);
-
-			return new CGSize(width, height);
-		}
-
-		/// <summary>
-		/// Gets the min value being left or right.
-		/// </summary>
-		/// <remarks>
-		/// This method kept here for readbility
-		/// of <see cref="SizeThatFits(IFrameworkElement, CGSize)"/> the keep its
-		/// fluent aspect.
-		/// It also does not use the generic extension that may create an very
-		/// short lived <see cref="IConvertible"/> instance.
-		/// </remarks>
-		private static nfloat LocalMin(this nfloat left, nfloat right)
-		{
-			return NMath.Min(left, right);
-		}
-
-		/// <summary>
-		/// Gets the max value being left or right.
-		/// </summary>
-		/// <remarks>
-		/// This method kept here for readability
-		/// of <see cref="SizeThatFits(IFrameworkElement, CGSize)"/> the keep its
-		/// fluent aspect.
-		/// It also does not use the generic extension that may create an very
-		/// short lived <see cref="IConvertible"/> instance.
-		/// </remarks>
-		private static nfloat LocalMax(this nfloat left, nfloat right)
-		{
-			return NMath.Max(left, right);
-		}
-
 		private static IFrameworkElement ConvertFromStubToElement(this IFrameworkElement element, IFrameworkElement originalRootElement, string name)
 		{
 			var elementStub = element as ElementStub;
@@ -425,56 +322,6 @@ namespace Microsoft.UI.Xaml
 			}
 			return element;
 		}
-
-		private static nfloat NumberOrDefault(this nfloat number, nfloat defaultValue)
-		{
-			return nfloat.IsNaN(number)
-				? defaultValue
-				: number;
-		}
-
-		public static void MaybeOrNot<TInstance>(this TInstance instance, Action nonNullAction, Action nullAction)
-		{
-			// Analysis disable once CompareNonConstrainedGenericWithNull
-			if (instance != null)
-			{
-				nonNullAction.Invoke();
-			}
-			else
-			{
-				nullAction.Invoke();
-			}
-		}
-
-#if __ANDROID__
-		/// <summary>
-		/// Applies the framework element constraints like the size and max size, using an already measured view.
-		/// </summary>
-		/// <param name="view"></param>
-		public static void OnMeasureOverride<T>(T view)
-			where T : View, IFrameworkElement
-		{
-			var updated = IFrameworkElementHelper
-				.SizeThatFits(view, new _Size(view.MeasuredWidth, view.MeasuredHeight).PhysicalToLogicalPixels())
-				.LogicalToPhysicalPixels();
-
-			Microsoft.UI.Xaml.Controls.Layouter.SetMeasuredDimensions(view, (int)updated.Width, (int)updated.Height);
-		}
-
-		/// <summary>
-		/// Applies the framework element constraints like the size and max size, using the provided measured size.
-		/// </summary>
-		/// <param name="view"></param>
-		public static void OnMeasureOverride<T>(T view, _Size measuredSize)
-			where T : View, IFrameworkElement
-		{
-			var updated = IFrameworkElementHelper
-				.SizeThatFits(view, new _Size(measuredSize.Width, measuredSize.Height).PhysicalToLogicalPixels())
-				.LogicalToPhysicalPixels();
-
-			Microsoft.UI.Xaml.Controls.Layouter.SetMeasuredDimensions(view, (int)updated.Width, (int)updated.Height);
-		}
-#endif
 
 		/// <summary>
 		/// Base constraint reasoning for simple containers that always respect the stretch of their children.
