--- src/Uno.UI/UI/Xaml/ILayouterElement.iOSmacOS.cs
@@ -1,32 +1,0 @@
-﻿#nullable enable
-using System;
-using System.Diagnostics.CodeAnalysis;
-using System.Runtime.CompilerServices;
-using Windows.Foundation;
-using CoreGraphics;
-using Uno.UI;
-
-namespace Microsoft.UI.Xaml;
-
-internal partial interface ILayouterElement
-{
-	[MethodImpl(MethodImplOptions.AggressiveInlining)]
-	internal bool XamlMeasureInternal(
-		Size availableSize,
-		Size? lastAvailableSize,
-		out CGSize measuredSize)
-	{
-		if (IsMeasureDirty || availableSize != lastAvailableSize)
-		{
-			if (this.DoMeasure(availableSize, out var measuredSizeLogical))
-			{
-				measuredSize = measuredSizeLogical.LogicalToPhysicalPixels();
-				return true;
-			}
-		}
-
-		// No need to measure
-		measuredSize = default;
-		return false;
-	}
-}
