+++ src/Uno.UI/UI/Xaml/Internal/RootVisual.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using Uno.UI.Extensions;
 using Windows.Foundation;
 using Windows.UI;
@@ -104,4 +105,32 @@ internal partial class RootVisual : Panel, IRootElement
 
 		return finalSize;
 	}
+
+#if __ANDROID__ || __IOS__ || __MACOS__
+	[MethodImpl(MethodImplOptions.AggressiveInlining)]
+	private void UpdateWindowContentLayout()
+	{
+		if (XamlRoot?.HostWindow?.Content is { } content)
+		{
+			content.UpdateLayout();
+		}
+	}
+
+#if __ANDROID__
+	protected override void OnLayoutCore(bool changed, int left, int top, int right, int bottom, bool localIsLayoutRequested)
+	{
+		base.OnLayoutCore(changed, left, top, right, bottom, localIsLayoutRequested);
+#elif __IOS__
+	public override void LayoutSubviews()
+	{
+		base.LayoutSubviews();
+#else
+	public override void Layout()
+	{
+		base.Layout();
+#endif
+		UpdateWindowContentLayout();
+	}
+
+#endif
 }
