+++ src/Uno.UI/UI/Xaml/UIElement.Layout.Flags.cs
@@ -1,0 +1,216 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using Uno.UI.Extensions;
+
+namespace Microsoft.UI.Xaml
+{
+	partial class UIElement
+	{
+		/// <summary>
+		/// Determines if InvalidateMeasure has been called
+		/// </summary>
+		internal bool IsMeasureDirty
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.MeasureDirty);
+		}
+
+		internal bool IsMeasureDirtyPath
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.MeasureDirtyPath);
+		}
+
+		internal bool IsMeasureDirtyOrMeasureDirtyPath
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsAnyLayoutFlagsSet(LayoutFlag.MeasureDirty | LayoutFlag.MeasureDirtyPath);
+		}
+
+		/// <summary>
+		/// If the first measure has been done since the control
+		/// is connected to its parent
+		/// </summary>
+		internal bool IsFirstMeasureDone
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.FirstMeasureDone);
+		}
+
+		/// <summary>
+		/// Determines if InvalidateArrange has been called
+		/// </summary>
+		internal bool IsArrangeDirty
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.ArrangeDirty);
+		}
+
+		internal bool IsArrangeDirtyPath
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsAnyLayoutFlagsSet(LayoutFlag.ArrangeDirtyPath);
+		}
+
+		internal bool IsArrangeDirtyOrArrangeDirtyPath
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsAnyLayoutFlagsSet(LayoutFlag.ArrangeDirty | LayoutFlag.ArrangeDirtyPath);
+		}
+
+		/// <summary>
+		/// If the first arrange has been done since the control
+		/// is connected to its parent
+		/// </summary>
+		internal bool IsFirstArrangeDone
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.FirstArrangeDone);
+		}
+
+		[Flags]
+		internal enum LayoutFlag : short
+		{
+			/// <summary>
+			/// Means the Measure is dirty for the current element
+			/// </summary>
+			MeasureDirty = 0b0000_0001,
+
+			/// <summary>
+			/// Means the Measure is dirty on at least one child of this element
+			/// </summary>
+			MeasureDirtyPath = 0b0000_0010,
+
+			/// <summary>
+			/// Indicates that the element is currently being measured during the Arrange cycle.
+			/// </summary>
+			MeasureDuringArrange = 0b0000_0001_0000_0000,
+
+			/// <summary>
+			/// Indicates that the element is currently being measured.
+			/// </summary>
+			MeasuringSelf = 0b0000_0010_0000_0000,
+
+			/// <summary>
+			/// Indicates that first measure has been done on the element after been connected to parent
+			/// </summary>
+			FirstMeasureDone = 0b0000_0100,
+
+			/// <summary>
+			/// Means the MeasureDirtyPath is disabled on this element.
+			/// </summary>
+			/// <remarks>
+			/// This flag is copied to children when they are attached, but can be re-enabled afterwards.
+			/// This flag is used during invalidation
+			/// </remarks>
+			MeasureDirtyPathDisabled = 0b0000_1000,
+
+			/// <summary>
+			/// Means the Arrange is dirty on the current element or one of its child
+			/// </summary>
+			ArrangeDirty = 0b0001_0000,
+
+			/// <summary>
+			/// Means the Arrange is dirty on at least one child of this element
+			/// </summary>
+			ArrangeDirtyPath = 0b0010_0000,
+
+			/// <summary>
+			/// Indicates that first arrange has been done on the element and we can use the
+			/// LayoutInformation.GetLayoutSlot() to get previous finalRect
+			/// </summary>
+			FirstArrangeDone = 0b0100_0000,
+
+			/// <summary>
+			/// Means the MeasureDirtyPath is disabled on this element.
+			/// </summary>
+			/// <remarks>
+			/// This flag is copied to children when they are attached, but can be re-enabled afterwards.
+			/// This flag is used during invalidation
+			/// </remarks>
+			ArrangeDirtyPathDisabled = 0b1000_0000,
+		}
+
+		private const LayoutFlag DEFAULT_STARTING_LAYOUTFLAGS = 0;
+		private const LayoutFlag LAYOUT_FLAGS_TO_CLEAR_ON_RESET =
+			LayoutFlag.MeasureDirty |
+			LayoutFlag.MeasureDirtyPath |
+			LayoutFlag.MeasureDuringArrange |
+			LayoutFlag.MeasuringSelf |
+			LayoutFlag.ArrangeDirty |
+			LayoutFlag.ArrangeDirtyPath |
+			LayoutFlag.FirstArrangeDone |
+			LayoutFlag.FirstMeasureDone;
+
+		private LayoutFlag _layoutFlags = DEFAULT_STARTING_LAYOUTFLAGS;
+
+		/// <summary>
+		/// Check for one specific layout flag
+		/// </summary>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal bool IsLayoutFlagSet(LayoutFlag flag) => (_layoutFlags & flag) == flag;
+
+		/// <summary>
+		/// Check that at least one of the specified flags is set
+		/// </summary>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal bool IsAnyLayoutFlagsSet(LayoutFlag flags) => (_layoutFlags & flags) != 0;
+
+		/// <summary>
+		/// Set one or many flags (set to 1)
+		/// </summary>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal void SetLayoutFlags(LayoutFlag flags)
+		{
+			global::System.Diagnostics.Debug.WriteLine($"Setting {flags} on {this},{this.GetDebugName()}");
+			_layoutFlags |= flags;
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal void SetLayoutFlags(LayoutFlag flags, bool state)
+		{
+			if (state)
+			{
+				SetLayoutFlags(flags);
+			}
+			else
+			{
+				ClearLayoutFlags(flags);
+			}
+		}
+
+		/// <summary>
+		/// Reset one or many flags (set flag to zero)
+		/// </summary>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal void ClearLayoutFlags(LayoutFlag flags)
+		{
+			_layoutFlags &= ~flags;
+			global::System.Diagnostics.Debug.WriteLine($"Clearing {flags} on {this},{this.GetDebugName()}");
+		}
+
+		/// <summary>
+		/// Reset flags to original state
+		/// </summary>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal void ResetLayoutFlags() => ClearLayoutFlags(LAYOUT_FLAGS_TO_CLEAR_ON_RESET);
+
+		internal bool IsMeasureDirtyPathDisabled
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.MeasureDirtyPathDisabled);
+
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			set => SetLayoutFlags(LayoutFlag.MeasureDirtyPathDisabled, value);
+		}
+
+		internal bool IsArrangeDirtyPathDisabled
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			get => IsLayoutFlagSet(LayoutFlag.ArrangeDirtyPathDisabled);
+
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			set => SetLayoutFlags(LayoutFlag.ArrangeDirtyPathDisabled, value);
+		}
+	}
+}
