+++ src/Uno.UI/UI/Xaml/UIElement.Layout.cs
@@ -1,255 +1,596 @@
-﻿#if __WASM__ || __SKIA__
-// "Managed Measure Dirty Path" means it's the responsibility of the
-// managed code (Uno) to walk the tree and do the measure phase.
-#define IMPLEMENTS_MANAGED_MEASURE_DIRTY_PATH
-#define IMPLEMENTS_MANAGED_ARRANGE_DIRTY_PATH
-#endif
+﻿#if !__NETSTD_REFERENCE__
 using System;
+using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Controls.Primitives;
+using Uno.Foundation.Logging;
+using Uno.UI;
+using Uno.UI.Extensions;
+using Uno.UI.Xaml;
+using Windows.Foundation;
+
+#if __ANDROID__
+using View = Android.Views.View;
+#elif __IOS__
+using UIKit;
+using View = UIKit.UIView;
+#elif __MACOS__
+using AppKit;
+using View = AppKit.NSView;
+#else
+using View = Microsoft.UI.Xaml.UIElement;
+#endif
 
 namespace Microsoft.UI.Xaml
 {
-	partial class UIElement
+	public partial class UIElement : DependencyObject
 	{
 		/// <summary>
-		/// Determines if InvalidateMeasure has been called
+		/// When set, measure and invalidate requests will not be propagated further up the visual tree, ie they won't trigger a re-layout.
+		/// Used where repeated unnecessary measure/arrange passes would be unacceptable for performance (eg scrolling in a list).
 		/// </summary>
-		internal bool IsMeasureDirty
+		internal bool ShouldInterceptInvalidate { get; set; }
+
+		public void InvalidateMeasure()
 		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.MeasureDirty);
-		}
-
-#if IMPLEMENTS_MANAGED_MEASURE_DIRTY_PATH
-		internal bool IsMeasureDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.MeasureDirtyPath);
-		}
-#else
-		// IsMeasureDirtyPath implemented in platform-specific file to
-		// connect to native mechanisms.
-#endif
-
-#if IMPLEMENTS_MANAGED_MEASURE_DIRTY_PATH
-		internal bool IsMeasureDirtyOrMeasureDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsAnyLayoutFlagsSet(LayoutFlag.MeasureDirty | LayoutFlag.MeasureDirtyPath);
-		}
-#else
-		/// <summary>
-		/// This is for compatibility - not implemented yet on this platform
-		/// </summary>
-		internal bool IsMeasureDirtyOrMeasureDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsMeasureDirty || IsMeasureDirtyPath;
-		}
-#endif
-
-		/// <summary>
-		/// If the first measure has been done since the control
-		/// is connected to its parent
-		/// </summary>
-		internal bool IsFirstMeasureDone
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.FirstMeasureDone);
-		}
-
-#if !__ANDROID__
-		/// <summary>
-		/// Determines if InvalidateArrange has been called
-		/// </summary>
-		internal bool IsArrangeDirty
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.ArrangeDirty);
-		}
-#endif
-
-#if IMPLEMENTS_MANAGED_ARRANGE_DIRTY_PATH
-		internal bool IsArrangeDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsAnyLayoutFlagsSet(LayoutFlag.ArrangeDirtyPath);
-		}
-
-		internal bool IsArrangeDirtyOrArrangeDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsAnyLayoutFlagsSet(LayoutFlag.ArrangeDirty | LayoutFlag.ArrangeDirtyPath);
-		}
-
-		/// <summary>
-		/// If the first arrange has been done since the control
-		/// is connected to its parent
-		/// </summary>
-		internal bool IsFirstArrangeDone
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.FirstArrangeDone);
-		}
-
-#else
-		/// <summary>
-		/// This is for compatibility - not implemented yet on this platform
-		/// </summary>
-		internal bool IsArrangeDirtyOrArrangeDirtyPath
-		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsArrangeDirty || IsArrangeDirtyPath;
-		}
-#endif
-
-		[Flags]
-		internal enum LayoutFlag : short
-		{
-			/// <summary>
-			/// Means the Measure is dirty for the current element
-			/// </summary>
-			MeasureDirty = 0b0000_0001,
-
-#if IMPLEMENTS_MANAGED_MEASURE_DIRTY_PATH
-			/// <summary>
-			/// Means the Measure is dirty on at least one child of this element
-			/// </summary>
-			MeasureDirtyPath = 0b0000_0010,
-
-			/// <summary>
-			/// Indicates that the element is currently being measured during the Arrange cycle.
-			/// </summary>
-			MeasureDuringArrange = 0b0000_0001_0000_0000,
-
-			/// <summary>
-			/// Indicates that the element is currently being measured.
-			/// </summary>
-			MeasuringSelf = 0b0000_0010_0000_0000,
-#endif
-
-			/// <summary>
-			/// Indicates that first measure has been done on the element after been connected to parent
-			/// </summary>
-			FirstMeasureDone = 0b0000_0100,
-
-			/// <summary>
-			/// Means the MeasureDirtyPath is disabled on this element.
-			/// </summary>
-			/// <remarks>
-			/// This flag is copied to children when they are attached, but can be re-enabled afterwards.
-			/// This flag is used during invalidation
-			/// </remarks>
-			MeasureDirtyPathDisabled = 0b0000_1000,
-
-#if !__ANDROID__ // On Android, it's directly connected to IsLayoutRequested
-			/// <summary>
-			/// Means the Arrange is dirty on the current element or one of its child
-			/// </summary>
-			ArrangeDirty = 0b0001_0000,
-#endif
-
-#if IMPLEMENTS_MANAGED_ARRANGE_DIRTY_PATH
-			/// <summary>
-			/// Means the Arrange is dirty on at least one child of this element
-			/// </summary>
-			ArrangeDirtyPath = 0b0010_0000,
-
-			/// <summary>
-			/// Indicates that first arrange has been done on the element and we can use the
-			/// LayoutInformation.GetLayoutSlot() to get previous finalRect
-			/// </summary>
-			FirstArrangeDone = 0b0100_0000,
-#endif
-
-			/// <summary>
-			/// Means the MeasureDirtyPath is disabled on this element.
-			/// </summary>
-			/// <remarks>
-			/// This flag is copied to children when they are attached, but can be re-enabled afterwards.
-			/// This flag is used during invalidation
-			/// </remarks>
-			ArrangeDirtyPathDisabled = 0b1000_0000,
-		}
-
-		private const LayoutFlag DEFAULT_STARTING_LAYOUTFLAGS = 0;
-		private const LayoutFlag LAYOUT_FLAGS_TO_CLEAR_ON_RESET =
-			LayoutFlag.MeasureDirty |
-#if IMPLEMENTS_MANAGED_MEASURE_DIRTY_PATH
-			LayoutFlag.MeasureDirtyPath |
-			LayoutFlag.MeasureDuringArrange |
-			LayoutFlag.MeasuringSelf |
-#endif
-#if !__ANDROID__
-			LayoutFlag.ArrangeDirty |
-#endif
-#if IMPLEMENTS_MANAGED_ARRANGE_DIRTY_PATH
-			LayoutFlag.ArrangeDirtyPath |
-			LayoutFlag.FirstArrangeDone |
-#endif
-			LayoutFlag.FirstMeasureDone;
-
-		private LayoutFlag _layoutFlags = DEFAULT_STARTING_LAYOUTFLAGS;
-
-		/// <summary>
-		/// Check for one specific layout flag
-		/// </summary>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal bool IsLayoutFlagSet(LayoutFlag flag) => (_layoutFlags & flag) == flag;
-
-		/// <summary>
-		/// Check that at least one of the specified flags is set
-		/// </summary>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal bool IsAnyLayoutFlagsSet(LayoutFlag flags) => (_layoutFlags & flags) != 0;
-
-		/// <summary>
-		/// Set one or many flags (set to 1)
-		/// </summary>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal void SetLayoutFlags(LayoutFlag flags) => _layoutFlags |= flags;
-
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal void SetLayoutFlags(LayoutFlag flags, bool state)
-		{
-			if (state)
+			if (ShouldInterceptInvalidate || IsMeasureDirty || IsLayoutFlagSet(LayoutFlag.MeasuringSelf))
 			{
-				SetLayoutFlags(flags);
+				return;
+			}
+
+			if (_traceLayoutCycle && this.Log().IsEnabled(LogLevel.Warning))
+			{
+				this.Log().LogWarning($"[LayoutCycleTracing] InvalidateMeasure {this},{this.GetDebugName()}");
+			}
+
+			SetLayoutFlags(LayoutFlag.MeasureDirty);
+
+			if (FeatureConfiguration.UIElement.UseInvalidateMeasurePath && !IsMeasureDirtyPathDisabled)
+			{
+				InvalidateParentMeasureDirtyPath();
 			}
 			else
 			{
-				ClearLayoutFlags(flags);
+				(this.GetParent() as UIElement)?.InvalidateMeasure();
+				if (IsVisualTreeRoot)
+				{
+					XamlRoot.InvalidateMeasure();
+				}
+			}
+
+#if __ANDROID__ || __IOS__ || __MACOS__
+			OnInvalidateMeasure();
+#endif
+		}
+
+#if __ANDROID__ || __IOS__ || __MACOS__
+		protected internal virtual void OnInvalidateMeasure()
+		{
+		}
+#endif
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void InvalidateMeasureDirtyPath()
+		{
+			if (IsMeasureDirtyOrMeasureDirtyPath)
+			{
+				return; // Already invalidated
+			}
+
+			SetLayoutFlags(LayoutFlag.MeasureDirtyPath);
+
+			InvalidateParentMeasureDirtyPath();
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		internal void InvalidateParentMeasureDirtyPath()
+		{
+			if (this.GetParent() is UIElement parent) //TODO: Should this use VisualTree.GetParent as fallback? https://github.com/unoplatform/uno/issues/8978
+			{
+				parent.InvalidateMeasureDirtyPath();
+			}
+#if __ANDROID__ || __IOS__ || __MACOS__
+			else if (!IsVisualTreeRoot && this.FindParent() is { } nativeParent)
+			{
+				while (nativeParent is not null && nativeParent is not UIElement)
+				{
+					LayoutInformation.SetMeasureDirtyPath(nativeParent, true);
+					nativeParent = nativeParent.FindParent();
+				}
+
+				if (nativeParent is UIElement uiElement)
+				{
+					uiElement.InvalidateMeasureDirtyPath();
+				}
+			}
+#endif
+			else if (IsVisualTreeRoot)
+			{
+				XamlRoot.InvalidateMeasure();
 			}
 		}
 
-		/// <summary>
-		/// Reset one or many flags (set flag to zero)
-		/// </summary>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal void ClearLayoutFlags(LayoutFlag flags) => _layoutFlags &= ~flags;
-
-		/// <summary>
-		/// Reset flags to original state
-		/// </summary>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal void ResetLayoutFlags() => ClearLayoutFlags(LAYOUT_FLAGS_TO_CLEAR_ON_RESET);
-
-		internal bool IsMeasureDirtyPathDisabled
+		public void InvalidateArrange()
 		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.MeasureDirtyPathDisabled);
+			if (ShouldInterceptInvalidate)
+			{
+				return;
+			}
 
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			set => SetLayoutFlags(LayoutFlag.MeasureDirtyPathDisabled, value);
+			if (IsArrangeDirty)
+			{
+				return; // Already dirty
+			}
+
+			if (_traceLayoutCycle)
+			{
+				if (this.Log().IsEnabled(LogLevel.Error))
+				{
+					this.Log().LogError($"[LayoutCycleTracing] InvalidateArrange {this},{this.GetDebugName()}");
+				}
+
+				if (this.Log().IsEnabled(LogLevel.Trace))
+				{
+					this.Log().Trace($"[LayoutCycleTracing] {Environment.StackTrace}");
+				}
+			}
+
+			SetLayoutFlags(LayoutFlag.ArrangeDirty);
+
+			if (FeatureConfiguration.UIElement.UseInvalidateArrangePath && !IsArrangeDirtyPathDisabled)
+			{
+				InvalidateParentArrangeDirtyPath();
+			}
+			else
+			{
+				(this.GetParent() as UIElement)?.InvalidateArrange();
+				if (IsVisualTreeRoot)
+				{
+					XamlRoot.InvalidateArrange();
+				}
+			}
 		}
 
-		internal bool IsArrangeDirtyPathDisabled
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void InvalidateArrangeDirtyPath()
 		{
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			get => IsLayoutFlagSet(LayoutFlag.ArrangeDirtyPathDisabled);
+			if (IsArrangeDirtyOrArrangeDirtyPath)
+			{
+				return; // Already invalidated
+			}
 
-			[MethodImpl(MethodImplOptions.AggressiveInlining)]
-			set => SetLayoutFlags(LayoutFlag.ArrangeDirtyPathDisabled, value);
+			SetLayoutFlags(LayoutFlag.ArrangeDirtyPath);
+
+			InvalidateParentArrangeDirtyPath();
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void InvalidateParentArrangeDirtyPath()
+		{
+			if (this.GetParent() is UIElement parent) //TODO: Should this use VisualTree.GetParent as fallback? https://github.com/unoplatform/uno/issues/8978
+			{
+				parent.InvalidateArrangeDirtyPath();
+			}
+#if __ANDROID__ || __IOS__ || __MACOS__
+			else if (!IsVisualTreeRoot && this.FindParent() is { } nativeParent && this is not PopupRoot)
+			{
+				while (nativeParent is not null && nativeParent is not UIElement)
+				{
+					LayoutInformation.SetArrangeDirtyPath(nativeParent, true);
+					nativeParent = nativeParent.FindParent();
+				}
+
+				if (nativeParent is UIElement uiElement)
+				{
+					uiElement.InvalidateArrangeDirtyPath();
+				}
+			}
+#endif
+			else //TODO: Why not check IsVisualTreeRoot as in InvalidateParentMeasureDirtyPath?
+			{
+				XamlRoot?.InvalidateArrange();
+			}
+		}
+
+		public void Measure(Size availableSize)
+		{
+#if __CROSSRUNTIME__
+			if (!_isFrameworkElement)
+#else
+			if (this is not FrameworkElement)
+#endif
+			{
+				return; // Only FrameworkElements are measurable
+			}
+
+			// Visibility should not be checked here. Consider the following scenario:
+			// 1. A collapsed element is measured before it enters the visual tree
+			// 2. Visibility changes to Visible after it enters the visual tree, which will call InvalidateMeasure
+			// In this case, we want step 1 to clear the dirty flag.
+			// If the flag isn't cleared in step 1, then InvalidateMeasure call in step 2 will do nothing because the
+			// element is already dirty, which means the dirtiness isn't propagated up to RootVisual.
+			// So, we want to go into DoMeasure, which will clear the flag.
+			// Then, DoMeasure is going to early return if Visibility is collapsed.
+
+			if (IsVisualTreeRoot)
+			{
+				MeasureVisualTreeRoot(availableSize);
+			}
+			else
+			{
+				// If possible we avoid the try/finally which might be costly on some platforms
+				DoMeasure(availableSize);
+			}
+		}
+
+		/// <remarks>
+		/// This method contains or is called by a try/catch containing method and
+		/// can be significantly slower than other methods as a result on WebAssembly.
+		/// See https://github.com/dotnet/runtime/issues/56309
+		/// </remarks>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void MeasureVisualTreeRoot(Size availableSize)
+		{
+			try
+			{
+				_isLayoutingVisualTreeRoot = true;
+				DoMeasure(availableSize);
+			}
+			finally
+			{
+				_isLayoutingVisualTreeRoot = false;
+			}
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void DoMeasure(Size availableSize)
+		{
+			var isFirstMeasure = !IsLayoutFlagSet(LayoutFlag.FirstMeasureDone);
+
+			var isDirty =
+				isFirstMeasure
+				|| (availableSize != m_previousAvailableSize)
+				|| IsMeasureDirty
+				|| !FeatureConfiguration.UIElement.UseInvalidateMeasurePath // dirty_path disabled globally
+				|| IsMeasureDirtyPathDisabled;
+
+			var isMeasureDirtyPath = IsMeasureDirtyPath;
+
+			if (!isDirty && !isMeasureDirtyPath)
+			{
+				return; // Nothing to do
+			}
+
+			if (isFirstMeasure)
+			{
+				SetLayoutFlags(LayoutFlag.FirstMeasureDone);
+			}
+
+			var remainingTries = MaxLayoutIterations;
+
+			while (--remainingTries > 0)
+			{
+				if (isDirty)
+				{
+					// We must reset the flag **BEFORE** doing the actual measure, so the elements are able to re-invalidate themselves
+					// TODO: This doesn't actually follow WinUI. It looks like in WinUI, the method
+					// CUIElement::MeasureInternal is doing SetIsMeasureDirty(FALSE); at the end.
+					// If we were able to align this to WinUI, we should remember clearing
+					// the flag in the Visibility == Visibility.Collapsed case as well.
+					// The Visibility condition is similar to the GetIsLayoutSuspended check in
+					// WinUI, which does goto Cleanup and will call SetIsMeasureDirty(FALSE);
+					ClearLayoutFlags(LayoutFlag.MeasureDirty | LayoutFlag.MeasureDirtyPath);
+
+					var prevSize = DesiredSize;
+
+					// The dirty flag is explicitly set on this element
+#if DEBUG
+					try
+#endif
+					{
+						if (this.Visibility == Visibility.Collapsed)
+						{
+							m_desiredSize = default;
+							RecursivelyApplyTemplateWorkaround();
+							return;
+						}
+
+						SetLayoutFlags(LayoutFlag.MeasuringSelf);
+						MeasureCore(availableSize);
+						ClearLayoutFlags(LayoutFlag.MeasuringSelf);
+						InvalidateArrange();
+					}
+#if DEBUG
+					catch (Exception ex)
+					{
+#if __CROSSRUNTIME__
+						_log.Error($"Error measuring {this}", ex);
+#else
+						_ = ex; // TODO
+#endif
+						throw;
+					}
+					finally
+#endif
+					{
+						m_previousAvailableSize = availableSize;
+
+						// if (!GetIsMeasureDuringArrange() && ! IsSameSize(prevSize, desiredSize) && !bInLayoutTransition)
+						if (!IsLayoutFlagSet(LayoutFlag.MeasureDuringArrange) && prevSize != DesiredSize)
+						{
+							if (GetUIElementAdjustedParentInternal() is { } pParent)
+							{
+								pParent.OnChildDesiredSizeChanged(pParent);
+							}
+						}
+					}
+
+					break;
+				}
+
+				// isMeasureDirtyPath is always true here
+				ClearLayoutFlags(LayoutFlag.MeasureDirtyPath);
+
+				// The dirty flag is set on one of the descendents:
+				// it will bypass the current element's MeasureOverride()
+				// since it shouldn't produce a different result and it's
+				// just a waste of precious CPU time to call it.
+				var children = this.GetChildren().GetEnumerator();
+
+				//foreach (var child in children)
+				while (children.MoveNext())
+				{
+					if (children.Current is UIElement { IsMeasureDirtyOrMeasureDirtyPath: true } child)
+					{
+						// If the child is dirty (or is a path to a dirty descendant child),
+						// We're remeasuring it.
+
+						var previousDesiredSize = child.DesiredSize;
+						child.EnsureLayoutStorage();
+						child.Measure(child.m_previousAvailableSize);
+						if (child.DesiredSize != previousDesiredSize)
+						{
+							isDirty = true;
+							break;
+						}
+					}
+#if __ANDROID__ || __IOS__ || __MACOS__
+					else if (children.Current is View nativeView && LayoutInformation.GetMeasureDirtyPath(nativeView))
+					{
+						var previousDesiredSize = LayoutInformation.GetDesiredSize(nativeView);
+						var newDesiredSize = MobileLayoutingHelpers.MeasureElement(nativeView, LayoutInformation.GetAvailableSize(nativeView));
+						if (newDesiredSize != previousDesiredSize)
+						{
+							isDirty = true;
+							break;
+						}
+					}
+#endif
+				}
+
+				children.Dispose(); // no "using" operator here to prevent an implicit try-catch on Wasm
+
+				if (isDirty)
+				{
+					continue;
+				}
+
+				break;
+			}
+		}
+
+		internal virtual void MeasureCore(Size availableSize)
+		{
+			throw new NotSupportedException("UIElement doesn't implement MeasureCore. Inherit from FrameworkElement, which properly implements MeasureCore.");
+		}
+
+		internal bool ShouldApplyLayoutClipAsAncestorClip()
+		{
+			return this is Panel; // Restrict to Panels, to limit app-compat risk
+								  //&& !GetIsScrollViewerHeader(); // Special-case:  ScrollViewer Headers, which can zoom, must scale the LayoutClip too
+		}
+
+		private void RecursivelyApplyTemplateWorkaround()
+		{
+			// Uno workaround. The template should NOT be applied here.
+			// But, without this workaround, VerifyVisibilityChangeUpdatesCommandBarVisualState test will fail.
+			// The real root cause for the test failure is that FindParentCommandBarForElement will
+			// return null, that is because Uno doesn't yet properly have a "logical parent" concept.
+			// We eagerly apply the template so that FindParentCommandBarForElement will
+			// find the command bar through TemplatedParent
+			if (this is Control thisAsControl)
+			{
+				thisAsControl.TryCallOnApplyTemplate();
+
+				// Update bindings to ensure resources defined
+				// in visual parents get applied.
+				this.UpdateResourceBindings();
+			}
+
+#if __CROSSRUNTIME__
+			foreach (var child in _children)
+#else
+			foreach (var childView in this.GetChildren())
+#endif
+			{
+#if !__CROSSRUNTIME__
+				if (childView is UIElement child)
+#endif
+				{
+					child.RecursivelyApplyTemplateWorkaround();
+				}
+			}
+		}
+
+		public void Arrange(Rect finalRect)
+		{
+#if __CROSSRUNTIME__
+			if (!_isFrameworkElement)
+#else
+			if (this is not FrameworkElement)
+#endif
+			{
+				return;
+			}
+
+			var firstArrangeDone = IsFirstArrangeDone;
+
+			if (Visibility == Visibility.Collapsed)
+			{
+				m_finalRect = finalRect;
+				HideVisual();
+				ClearLayoutFlags(LayoutFlag.ArrangeDirty | LayoutFlag.ArrangeDirtyPath);
+				return;
+			}
+
+			if (firstArrangeDone && !IsArrangeDirtyOrArrangeDirtyPath && finalRect == m_finalRect)
+			{
+				ClearLayoutFlags(LayoutFlag.ArrangeDirty | LayoutFlag.ArrangeDirtyPath);
+				return; // Calling Arrange would be a waste of CPU time here.
+			}
+
+			if (IsVisualTreeRoot)
+			{
+				ArrangeVisualTreeRoot(finalRect);
+			}
+			else
+			{
+				// If possible we avoid the try/finally which might be costly on some platforms
+				DoArrange(finalRect);
+			}
+		}
+
+		/// <remarks>
+		/// This method contains or is called by a try/catch containing method and can be significantly slower than other methods as a result on WebAssembly.
+		/// See https://github.com/dotnet/runtime/issues/56309
+		/// </remarks>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void ArrangeVisualTreeRoot(Rect finalRect)
+		{
+			try
+			{
+				_isLayoutingVisualTreeRoot = true;
+				DoArrange(finalRect);
+			}
+			finally
+			{
+				_isLayoutingVisualTreeRoot = false;
+			}
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		private void DoArrange(Rect finalRect)
+		{
+			var isFirstArrange = !IsLayoutFlagSet(LayoutFlag.FirstArrangeDone);
+
+			var isDirty =
+				isFirstArrange
+				|| IsArrangeDirty
+				|| finalRect != m_finalRect;
+
+			if (!isDirty && !IsArrangeDirtyPath)
+			{
+				return; // Nothing do to
+			}
+
+			if (GetUseLayoutRounding())
+			{
+				finalRect.X = LayoutRound(finalRect.X);
+				finalRect.Y = LayoutRound(finalRect.Y);
+				finalRect.Width = LayoutRound(finalRect.Width);
+				finalRect.Height = LayoutRound(finalRect.Height);
+			}
+
+			var remainingTries = MaxLayoutIterations;
+
+			while (--remainingTries > 0)
+			{
+				if (IsMeasureDirtyOrMeasureDirtyPath)
+				{
+					// Uno doc: in WinUI, the flag is only set and reset if IsMeasureDirty, not IsMeasureDirtyOrMeasureDirtyPath
+					SetLayoutFlags(LayoutFlag.MeasureDuringArrange);
+					DoMeasure(m_previousAvailableSize);
+					ClearLayoutFlags(LayoutFlag.MeasureDuringArrange);
+				}
+
+				if (isDirty)
+				{
+					ShowVisual();
+
+					// We must store the updated slot before natively arranging the element,
+					// so the updated value can be read by indirect code that is being invoked on arrange.
+					// For instance, the EffectiveViewPort computation reads that value to detect slot changes (cf. PropagateEffectiveViewportChange)
+					m_finalRect = finalRect;
+
+					// We must reset the flag **BEFORE** doing the actual arrange, so the elements are able to re-invalidate themselves
+					ClearLayoutFlags(LayoutFlag.ArrangeDirty | LayoutFlag.ArrangeDirtyPath);
+
+					ArrangeCore(finalRect);
+
+					SetLayoutFlags(LayoutFlag.FirstArrangeDone);
+
+					break;
+				}
+				else if (IsArrangeDirtyPath)
+				{
+					ClearLayoutFlags(LayoutFlag.ArrangeDirtyPath);
+
+					var children = this.GetChildren().GetEnumerator();
+
+					while (children.MoveNext())
+					{
+						var child = children.Current;
+
+						if (child is UIElement { IsArrangeDirtyOrArrangeDirtyPath: true } childAsUIElement)
+						{
+							var previousRenderSize = childAsUIElement.RenderSize;
+							childAsUIElement.Arrange(childAsUIElement.m_finalRect);
+
+							if (childAsUIElement.RenderSize != previousRenderSize)
+							{
+								isDirty = true;
+								break;
+							}
+						}
+#if __ANDROID__ || __IOS__ || __MACOS_
+						else if (child is View nativeView && LayoutInformation.GetArrangeDirtyPath(nativeView))
+						{
+							// Using LayoutSlot as an approximation for RenderSize, which is UIElement-concept
+							var previousRenderSize = LayoutInformation.GetLayoutSlot(child).Size;
+							MobileLayoutingHelpers.ArrangeElement(child, finalRect);
+
+							if (finalRect.Size != previousRenderSize)
+							{
+								isDirty = true;
+								break;
+							}
+						}
+#endif
+					}
+
+					children.Dispose(); // no "using" operator here to prevent an implicit try-catch on Wasm
+
+					if (!isDirty)
+					{
+						break;
+					}
+				}
+				else
+				{
+					break;
+				}
+			}
+
+		}
+
+		partial void HideVisual();
+		partial void ShowVisual();
+
+		internal virtual void ArrangeCore(Rect finalRect)
+		{
+			throw new NotSupportedException("UIElement doesn't implement ArrangeCore. Inherit from FrameworkElement, which properly implements ArrangeCore.");
 		}
 	}
 }
+#endif
